+++ src/hasp/hasp.cpp
@@ -91,7 +91,8 @@ lv_style_t style_mbox_bg; /*Black bg. style with opacity*/
 lv_obj_t* kb;
 // lv_font_t * defaultFont;
 
-static lv_font_t* haspFonts[12] = {nullptr};
+// static lv_font_t* haspFonts[12] = {nullptr};
+static lv_font_t* haspFonts[12] __NOINIT_ATTR;
 uint8_t current_page            = 1;
 
 /**