+++ src/sys/net/hasp_time.cpp
@@ -667,7 +667,7 @@ bool timeGetConfig(const JsonObject& settings)
     settings["ntp"][2] = preferences.getString("ntp3", NTPSERVER3);
     preferences.end();
 
-#if ESP_ARDUINO_VERSION_MAJOR >= 2
+#if ESP_ARDUINO_VERSION_MAJOR == 2
     nvs_iterator_t it = nvs_entry_find("config", "time", NVS_TYPE_ANY);
     while(it != NULL) {
         nvs_entry_info_t info;
@@ -675,6 +675,9 @@ bool timeGetConfig(const JsonObject& settings)
         it = nvs_entry_next(it);
         printf("key '%s', type '%d' \n", info.key, info.type);
     };
+    if (it != NULL) {
+        nvs_release_iterator(it);
+    }
 #endif
 
     if(changed) configOutput(settings, TAG_TIME);