+++ tools/analyze_elf.py
@@ -7,7 +7,7 @@ Import("env")
 env.AddTarget(
     name = "analyze_ram",
     dependencies = "$BUILD_DIR\${PROGNAME}${PROGSUFFIX}",
-    actions = 'python $PROJECT_DIR/tools/elf-size-analyze.py $BUILD_DIR\${PROGNAME}${PROGSUFFIX} -t $PROJECT_PACKAGES_DIR\\toolchain-xtensa32\\bin\\xtensa-esp32-elf- -a -H -w 120 -R -m 512',
+    actions = 'python $PROJECT_DIR/tools/elf-size-analyze.py $BUILD_DIR\${PROGNAME}${PROGSUFFIX} -t $PROJECT_PACKAGES_DIR\\toolchain-xtensa-esp32\\bin\\xtensa-esp32-elf- -a -H -w 120 -R -m 10240',
     title = "Analyze RAM",
     description = "Build and analyze",
     group="Analyze"
@@ -15,7 +15,7 @@ env.AddTarget(
 env.AddTarget(
     name = "analyze_flash",
     dependencies = "$BUILD_DIR\${PROGNAME}${PROGSUFFIX}",
-    actions = 'python $PROJECT_DIR/tools/elf-size-analyze.py $BUILD_DIR\${PROGNAME}${PROGSUFFIX} -t $PROJECT_PACKAGES_DIR\\toolchain-xtensa32\\bin\\xtensa-esp32-elf- -a -H -w 120 -F -m 512',
+    actions = 'python $PROJECT_DIR/tools/elf-size-analyze.py $BUILD_DIR\${PROGNAME}${PROGSUFFIX} -t $PROJECT_PACKAGES_DIR\\toolchain-xtensa-esp32\\bin\\xtensa-esp32-elf- -a -H -w 120 -F -m 10240',
     title = "Analyze Flash",
     description = "Build and analyze",
     group="Analyze"